Package the compression corpus into a zip archive.
Add a genrule to zip up all the corpora and depend just on that. At runtime, we read the entire thing into memory and cache it so we don't need to go to disk again.

This lets us delete the generate_data_fleet genrule and simplify the build.

We take a dependency on libzip to process the zip archive.

PiperOrigin-RevId: 624214938
Change-Id: I0c77daba228ac167eb3fbb022fc574522383c428