First draft of a tool for running fleetbench benchmarks in parallel.
This runs the specified set of benchmarks for a specified duration, trying to achieve the specified system CPU utilization.

Benchmarks are randomly selected for scheduling based on their duration, so that shorter-running benchmarks are more likely to be selected. We naively pick the least-loaded CPU to assign the job to.

The output is written to a temp directory for later processing.

PiperOrigin-RevId: 632609998
Change-Id: I04611b15a611fd5fff024c72735d2f53b2eaea78