Add support for different scheduling modes in parallel_bench
This CL adds support for different scheduling modes in parallel_bench. The scheduling mode is controlled by the `--scheduling_mode` flag, with default  mode as DYNAMIC. The following modes are supported:

* DISABLE: SMT is disabled.
* DYNAMIC: SMT is enabled and the scheduler is allowed to pin workloads.
* SKEWED: SMT is enabled and workloads are pinned to Socket0 and lower hyperthreads as primary and higher hyperthreads as secondary.
* BALANCED: SMT is enabled and workloads are balanced across sockets and hyperthreads.

This CL also adds a function to read CPU information from `/proc/cpuinfo` and organizes it into a dictionary. This dictionary structure is:

```
{ physical_id: { core_id: [vcpu_1, vcpu_2, ...] } }
```

This dictionary helps selecting a range of CPUs based on the provided scheduling mode.

PiperOrigin-RevId: 683295168
Change-Id: Iddf63fae831933f35c55101cccdb03a3daadc3aa