Add .bazelversion file.
The build currently fails like so with Bazel 8:

ERROR: Error computing the main repository mapping: error loading package 'external': Both --enable_bzlmod and --enable_workspace are disabled, but one of them must be enabled to fetch external dependencies.

Passing one of the suggested flags doesn't fix the build either. Fix
it by using the .bazelversion file to pin to a version that works.

PiperOrigin-RevId: 704812165
Change-Id: Id7f6f74769c237437f6cc4cd7f70c3580a7b404d