Fix InsertMiss_Cold
InsertMiss_Cold is supposed to insert keys that are not present in the set. However, currently, this is only the case in the first iteration of the outer loop. In the following iterations, it inserts the same keys as before which are now already present in the set.

This CL fixes this by removing each element again after insertion. The removals are performed while the benchmark timer is paused. Even though `PauseTiming()/ResumeTiming()` are relatively expensive, it is OK to use them here because `sets` is large, and thus the cost of the two functions is small compared to the cost of the inner loop.

PiperOrigin-RevId: 744835970
Change-Id: Ieca3052a2b92f24553be634b51197cc9332f2881