Create a copy of the sets vector in InsertMiss_Hot
The benchmark loop of the InsertMiss_Hot benchmark modifies the sets in the vector. Therefore, we need to make a copy. Otherwise, repeated calls to `InsertMiss_Hot` (e.g., if the `--benchmark_repetitions` flag is set to a value larger than 1) start with increasingly larger sets, and thus the runtime of the benchmark grows with each repetition.

PiperOrigin-RevId: 747418678
Change-Id: I3327d6bbd6b9cb148c45100985f21e6148c60ffd