Fix InsertMiss_Cold
`sc.GetNonExistingKeys` requires that the sets vector that is passed as a reference parameter is not modified between calls to this function. However, this requirement is currently not satisfied if `kLookup` is false. Because of this, runs of different benchmarks are not independent, and later benchmarks can erroneously get the keys vector of earlier benchmarks.

PiperOrigin-RevId: 748662999
Change-Id: I3a31276387625f2cbf889139acb8ad6e076221ae